requirements.txt
azure-storage-blob
toml